Somecode cahnges

+++ SolidDemo/LoanAccounts/Interfaces/ILoan.cs
@@ -7,7 +7,7 @@ public interface ILoan
     int LoanId { get; }
     decimal TotalAmount();
     decimal MonthlyAmortization();
-    int Duration { get; }
+    decimal Duration { get; }
     LoanType LoanType { get; }
     void OutputMessage();
 }
